// src/modules/validation/types.ts
export interface ValidationView {
    readonly id: number;
    readonly groupId: number;
    readonly createdBy: number;
    readonly cloneId: number;
    readonly cloneName: string;
    readonly proteinId: number;
    readonly proteinName: string;
    readonly lotId: number | null;
    readonly userId: number | null;
    readonly userName: string | null;
    readonly lotName: string | null;
    readonly lotNumber: string | null;
    readonly conjugateId: number | null;
    readonly tubeNumber: string | number;
    readonly speciesId: number | null;
    readonly speciesName: string | null;
    readonly application: number;
    readonly positiveControl: string;
    readonly negativeControl: string;
    readonly incubationConditions: string;
    readonly concentration: string;
    readonly concentrationUnit: string;
    readonly tissue: string;
    readonly fixation: number;
    readonly fixationNotes: string;
    readonly notes: string;
    readonly status: number;
    readonly antigenRetrievalType: string;
    readonly antigenRetrievalTime: string;
    readonly antigenRetrievalTemperature: string;
    readonly saponin: boolean | null;
    readonly saponinConcentration: string | null;
    readonly methanolTreatment: boolean | null;
    readonly methanolTreatmentConcentration: string | null;
    readonly surfaceStaining: boolean | null;
    readonly surfaceStainingConcentration: string | null;
    readonly meta: object;
    readonly isArchived: boolean;
    readonly createdAt: string;
    readonly updatedAt: string;
    readonly validationFiles: ValidationFileView[];
}